fix(util): add a newline to the last line of the text
When the hacker.nvim reaches the end of the available text,
it starts writing out the same text beneath the one it just wrote.
When that happens, there is no expected new line between the previous
text's last line and the new text's first line
(the two mentioned lines merge into one).